move Project::create_sample_computed_files to Sample model, move Project::process_computed_file to ComputedFile model
